Fix completion with dashes (#4)
* Fix completion with dashes

If I write something like curl --he<tab> it currently crashes on this
string match line. This should solve this by telling the string command
that it won't be any switches after this `--`-mark.

* Handle dashes in string and commandline

As the input might be using dashes we need to use double dashes to tell
the command that there won't be any more options.